Update dev dependencies
semistandard can be updated too, which fixes the circular
dependency warnings.

However this also changes the linting rules, including a change to
remove `var` in favour of `let` or `const`. This could potentially be a
breaking change for customers on older node versions, so we'll leave it
for a separate PR to be released under a major version number bump to
signify this change.